vector load/store support stack

+++ riscv/interactive.cc
@@ -347,7 +347,7 @@ void sim_t::interactive_vreg(const std::string& cmd, const std::vector<std::stri
 
   std::ostream out(sout_.rdbuf());
   out << std::dec << "VLEN=" << (vlen << 3) << " bits; ELEN=" << (elen << 3) << " bits" << std::endl;
-  for (reg_t vu_idx=0; vu_idx<static_cast<int>(n_vu); vu_idx++) {
+  for (reg_t vu_idx=0; vu_idx < n_vu; vu_idx++) {
     for (int r = rstart; r < rend; ++r) {
       out << std::setfill (' ') << std::left << std::setw(4) << vr_name[r] << std::right << ": ";
       for (int e = num_elem-1; e >= 0; --e){
